死亡日別による死亡者数の推移: Rename components/cards/DeathsByDeathDateCard.vue to components/index/CardsReference/DeathsByDeathDate/Card.vue
