+++ components/index/CardsReference/DeathsByDeathDate/Card.vue
